bugfixes, add dead player check

+++ start_player.sh
@@ -3,5 +3,7 @@
 python3 -m venv venv
 source venv/bin/activate
 
-python3 -m pip install -r requirements.txt
+if test -f "requirements.txt"; then
+  python3 -m pip install -r requirements.txt
+fi
 python3 main.py
